+++ src/api/corpusConfig.test.ts
@@ -92,6 +92,8 @@ describe("parseConfig", () => {
       metadata: {
         name: { swe: "Nyheter", eng: "News" },
         description: { swe: "Senaste nytt", eng: "Latest news" },
+        language: "swe",
+        variety: "fsv",
       },
       import: {
         importer: "xml_import:parse",
@@ -112,6 +114,8 @@ describe("parseConfig", () => {
       name: { swe: "Nyheter", eng: "News" },
       description: { swe: "Senaste nytt", eng: "Latest news" },
       textAnnotation: "article",
+      language: "swe",
+      variety: "fsv",
       sentenceSegmenter: "linebreaks",
       datetime: {
         from: "2000-01-01",