+++ src/corpus/views/CorpusConfigurationView.vue
@@ -5,7 +5,7 @@ import { useI18n } from "vue-i18n";
 import { useRouter } from "vue-router";
 import { FormKit } from "@formkit/vue";
 import { PhLightbulbFilament, PhTrash } from "@phosphor-icons/vue";
-import { watchImmediate } from "@vueuse/core";
+import { computedAsync, watchImmediate } from "@vueuse/core";
 import { omit, pickBy } from "es-toolkit";
 import { useCorpus } from "../corpus.composable";
 import {
@@ -23,7 +23,7 @@ import useResourceIdParam from "@/resource/resourceIdParam.composable";
 import RouteButton from "@/components/RouteButton.vue";
 import useAlert from "@/alert/alert.composable";
 import PendingContent from "@/spin/PendingContent.vue";
-import type { ByLang } from "@/util";
+import { fromKeys, type ByLang } from "@/util";
 import TerminalOutput from "@/components/TerminalOutput.vue";
 import useLocale from "@/i18n/locale.composable";
 import TabsBar from "@/components/TabsBar.vue";
@@ -32,7 +32,8 @@ import useSources from "@/resource/sources.composable";
 import { CORPUS_SOURCE_FORMATS } from "@/file";
 import { useUserStore } from "@/store/user.store";
 import { useSparvAnalyses } from "@/api/useSparvAnalyses";
-import type { SparvAnalysis } from "@/api/api.types";
+import { useApi } from "@/api/useApi";
+import { useAppConfig } from "@/app/useAppConfig";
 
 type TabKey = "metadata" | "settings" | "analyses";
 
@@ -41,17 +42,21 @@ type Form = {
   description: ByLang;
   format: CorpusSourceFormat;
   textAnnotation: string;
+  language: string;
   sentenceSegmenter: ConfigSentenceSegmenter;
   datetimeFrom: string;
   datetimeTo: string;
   analyses: Record<string, boolean>;
 };
 
+const { appConfig } = useAppConfig();
 const router = useRouter();
 const id = useResourceIdParam();
+const api = useApi();
 const { config, configOptions, saveConfigOptions } = useCorpus(id);
 const { extensions } = useSources("corpus", id);
-const { analyses, getAnalysesByAnnotations } = useSparvAnalyses();
+const { analyses, getAnalysesByAnnotations, getLanguageCode } =
+  useSparvAnalyses();
 const { showAlert } = useAlert();
 const { t } = useI18n();
 const { locale3, th, thCompare } = useLocale();
@@ -59,30 +64,58 @@ const { canAdmin, canWrite } = useUserStore();
 
 const tabSelected = ref<TabKey>("metadata");
 
-/** Selectable analyses grouped by unit */
-const analysesGrouped = computed(() => {
+const languages = computedAsync(() => api.sparvLanguages(), []);
+
+const languageOptions = computed(() =>
+  languages.value.map(({ code, name, variety }) => ({
+    value: getLanguageCode(code, variety),
+    label: name,
+  })),
+);
+
+const selectedLanguage = computed(() => {
+  if (!configOptions.value?.language) return appConfig.defaultLanguage;
+  return getLanguageCode(
+    configOptions.value.language,
+    configOptions.value.variety,
+  );
+});
+
+/** Analyses grouped by language and unit */
+const analysisGroups = computed(() => {
+  // Require attributive annotations
+  const filtered = analyses.value
+    .filter((analysis) =>
+      analysis.annotations.some((annotation) => annotation.includes(":")),
+    )
+    .sort(thCompare((x) => x.name));
+
+  // Language codes
+  const codes = languageOptions.value.map((l) => l.value);
   // List of recognized units
   const units = ["text", "sentence", "token", "other"];
 
-  function filter(unit: string) {
-    return analyses.value
-      .filter((analysis) => match(analysis, unit))
-      .sort(thCompare((x) => x.name));
-  }
-
-  function match(analysis: SparvAnalysis, requiredUnit: string) {
-    // Require attributive annotations
-    if (!analysis.annotations.some((annotation) => annotation.includes(":")))
-      return false;
-    // Recognize the analysis unit
-    const unit = units.includes(analysis.analysis_unit?.eng)
-      ? analysis.analysis_unit?.eng
-      : "other";
-    // Match the given unit
-    return unit == requiredUnit;
-  }
-
-  return Object.fromEntries(units.map((unit) => [unit, filter(unit)]));
+  // Build two-dimensional listing of analyses by language and unit
+  return fromKeys(codes, (code) =>
+    fromKeys(units, (unit) =>
+      filtered.filter((analysis) => {
+        // Match language
+        // Sparv handles `language_varieties` separately from `languages`,
+        // but we'll assume there is only one variety if any.
+        const variety = analysis.language_varieties?.[0];
+        const matchesLanguage =
+          !analysis.languages ||
+          analysis.languages.find(
+            (l) => getLanguageCode(l.code, variety) == code,
+          );
+        if (!matchesLanguage) return false;
+        // Match unit
+        const unitRaw = analysis.analysis_unit?.eng || "";
+        const thisUnit = units.includes(unitRaw) ? unitRaw : "other";
+        return thisUnit == unit;
+      }),
+    ),
+  );
 });
 
 const formatOptions = computed<FormKitOptionsList>(() =>
@@ -126,6 +159,9 @@ watchImmediate(configOptions, () => {
 async function submit(fields: Form) {
   const configOld = original.value;
 
+  // Split language-variety code
+  const [language, variety] = fields.language.split("-");
+
   // Use datetime if both are set
   const datetime =
     fields.datetimeFrom && fields.datetimeTo
@@ -139,6 +175,8 @@ async function submit(fields: Form) {
 
   const configNew: ConfigOptions = {
     ...omit(fields, ["datetimeFrom", "datetimeTo"]),
+    language,
+    variety,
     datetime,
     annotations,
   };
@@ -327,6 +365,17 @@ async function submit(fields: Form) {
                 </i18n-t>
               </HelpBox>
 
+              <FormKit
+                name="language"
+                :label="$t('config.language')"
+                type="select"
+                :value="selectedLanguage"
+                input-class="w-72"
+                :options="languageOptions"
+                validation="required"
+                :help="$t('config.language.help')"
+              />
+
               <FormKit type="group" name="analyses">
                 <table class="my-2">
                   <thead>
@@ -336,7 +385,12 @@ async function submit(fields: Form) {
                       <th>{{ $t("config.analyses.task") }}</th>
                     </tr>
                   </thead>
-                  <tbody v-for="(group, unit) in analysesGrouped" :key="unit">
+                  <tbody
+                    v-for="(group, unit) in analysisGroups[
+                      (value as Form).language
+                    ]"
+                    :key="unit"
+                  >
                     <tr>
                       <th colspan="5" class="text-lg pt-4!">
                         {{ $t("config.analyses.unit") }}: